Add tables names on entities, change Long to Integer in repositories and REST Controller and check if user already exist un User REST Controller and add email validation for User Model

+++ src/main/java/com/epitools/homer/homer/repository/ContributorRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContributorRepository extends JpaRepository<Contributor, Long> {}
+public interface ContributorRepository extends JpaRepository<Contributor, Integer> {}
